feature/#5: 좋아요 두 번 생성 방지 및 유효성 검사 추가
