--- .github/ISSUE_TEMPLATE/issue_template.md
@@ -1,18 +1,0 @@
----
-name: issue_template
-about: issue template
-title: ''
-labels: ''
-assignees: ''
-
----
-
-## 개요
-- 상품 게시판 작업을 위한 상품 공통 컴포넌트 작업 및 API 연동 진행
-
-## 작업 목록
-- [ ] 메인 페이지 UI 구현
-- [ ] 메인 페이지 API 연동
-
-## 참고
-- 작업하면서 참고한 레퍼런스 문서 등 추가
