--- post-love/.github/CODEOWNERS
@@ -1,1 +1,0 @@
-* @cyoure @jinwoo0207 @Contingency1
