--- post-love/app/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "app",
-  "version": "1.0.0",
-  "description": "mini poroject",
-  "main": "index.js",
-  "scripts": {
-    "start": "nodemon ./bin/www.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "modern agile 7",
-  "license": "MIT",
-  "dependencies": {
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.1",
-    "express": "^4.18.2",
-    "express-validator": "^7.0.1",
-    "mysql2": "^3.9.1",
-    "nodemon": "^3.0.3"
-  },
-  "type": "module"
-}
