--- post-love/app/src/config/db.js
@@ -1,14 +1,0 @@
-import mysql from "mysql2/promise";
-import * as dotenv from "dotenv";
-
-dotenv.config();
-
-const db = mysql.createPool({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_DATABASE,
-  timezone: "Asia/Seoul",
-});
-
-export default db;
